+++ linux/tools/SmartTofCli/fw_upgrade.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+./dmcam-cli -e "flash mcu smart_tof_e2_fw_20180521.bin"